https://github.com/Carlosmapego8/GoMountain/pull/12#discussion_r1872895450 Mueve todos los tipos a un mismo paquete
Todos los tipos se definen en el paquete gestorturnos.
Se modifica también la estructura de directorios, todos los tipos se encuentran en internal sin subdirectorios.
