+++ internal/auxiliar.go
@@ -1,0 +1,9 @@
+package gestorturnos
+
+type Auxiliar struct {
+	ID             int
+	Contrato       Contrato
+	Turnos         []Turno
+	TurnoPreferido TipoTurno
+	DiasBaja       int
+}
