+++ internal/gestorturnos.go
@@ -1,9 +1,5 @@
 package gestorturnos
 
-import (
-	"GoMountain/internal/auxiliar"
-)
-
 type GestorTurnos struct {
-	Auxiliares []auxiliar.Auxiliar
+	Auxiliares []Auxiliar
 }
